requirements.txt
alabaster==0.7.7
amqp==1.4.8
anyjson==0.3.3
appnope==0.1.0
awsebcli==3.7.6
Babel==2.1.1
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.4.1
billiard==3.3.0.22
blessed==1.9.5
botocore==1.4.25
brewer2mpl==1.4.1
cairocffi==0.7.2
celery==3.1.19
cement==2.8.2
certifi==2015.11.20.1
cffi==1.5.2
click==6.2
colorama==0.3.7
cycler==0.9.0
decorator==4.0.6
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
dropbox==5.2.2
feedparser==5.2.1
Flask==0.10.1
Flask-SocketIO==2.0
Flask-Sockets==0.1
Flask-SQLAlchemy==2.1
functools32==3.2.3.post2
gevent==1.0.2
gevent-websocket==0.9.5
ggplot==0.6.8
gnureadline==6.3.3
greenlet==0.4.9
gunicorn==19.4.1
httplib2==0.9.2
ipykernel==4.2.2
ipython==4.0.1
ipython-genutils==0.1.0
ipywidgets==4.1.1
iron-cache==0.3.2
iron-celery==0.4.5
iron-core==1.2.0
iron-mq==0.8
itsdangerous==0.24
Jinja2==2.7.3
jmespath==0.9.0
joblib==0.9.3
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.1.1
jupyter-console==4.1.0
jupyter-core==4.0.6
kombu==3.0.32
MarkupSafe==0.23
matplotlib==1.5.1
mechanize==0.2.5
memory-profiler==0.40
mistune==0.7.1
multiprocessing==2.6.2.1
nbconvert==4.1.0
nbformat==4.0.1
nltk==3.2.1
notebook==4.1.0
numpy==1.10.4
oauth2==1.9.0.post1
palettable==2.1.1
pandas==0.17.1
Paste==2.0.2
path.py==8.1.2
pathspec==0.3.4
patsy==0.4.1
pexpect==4.0.1
pickleshare==0.5
Pillow==3.0.0
protobuf==2.6.1
psutil==3.3.0
psycopg2==2.6.1
ptyprocess==0.5
pycparser==2.14
Pygments==2.0.2
pymc==2.3.6
pyparsing==2.1.0
pytagcloud==0.3.5
python-dateutil==2.5.3
python-engineio==0.8.6
python-socketio==0.9.1
pytz==2015.7
PyYAML==3.11
pyzmq==15.2.0
qtconsole==4.1.1
redis==2.10.5
requests==2.9.1
rq==0.5.6
scikit-learn==0.17.1
scipy==0.16.1
selenium==2.53.1
semantic-version==2.5.0
simplegeneric==0.8.1
simplejson==3.8.2
singledispatch==3.4.0.3
six==1.10.0
sklearn==0.0
snowballstemmer==1.2.1
Sphinx==1.3.3
sphinx-rtd-theme==0.1.9
SQLAlchemy==1.0.11
statsmodels==0.6.1
terminado==0.6
textblob==0.11.1
texttable==0.8.4
tornado==4.3
traitlets==4.0.0
urllib3==1.14
virtualenv==13.1.2
wcwidth==0.1.6
websocket-client==0.37.0
Werkzeug==0.10.4
wordcloud==1.2.1
xlrd==0.9.4
